Fix chart of trade view with contract
